Split workflow tests, so they can be parallelized on the CI for faster turnaround time.
