+++ BuildResidentialHPXML/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>build_residential_hpxml</name>
   <uid>a13a8983-2b01-4930-8af2-42030b6e4233</uid>
-  <version_id>d77dae30-70b5-4522-90fe-fb58f53cc600</version_id>
-  <version_modified>2023-11-01T02:43:51Z</version_modified>
+  <version_id>000ef907-6e7e-4dfc-adf1-346cc0f57933</version_id>
+  <version_modified>2023-11-02T14:39:56Z</version_modified>
   <xml_checksum>2C38F48B</xml_checksum>
   <class_name>BuildResidentialHPXML</class_name>
   <display_name>HPXML Builder</display_name>
@@ -6777,7 +6777,7 @@
       <checksum>84CB0D65</checksum>
     </file>
     <file>
-      <filename>build_residential_hpxml_test.rb</filename>
+      <filename>test_build_residential_hpxml.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
       <checksum>F24631AD</checksum>
