+++ BuildResidentialHPXML/tests/test_build_residential_hpxml.rb
