+++ BuildResidentialScheduleFile/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>build_residential_schedule_file</name>
   <uid>f770b2db-1a9f-4e99-99a7-7f3161a594b1</uid>
-  <version_id>2c589b36-e096-4d7b-bccd-3880fb9624fe</version_id>
-  <version_modified>2023-10-31T23:05:37Z</version_modified>
+  <version_id>48439a72-2880-4934-b216-38a24bf684b7</version_id>
+  <version_modified>2023-11-02T14:39:58Z</version_modified>
   <xml_checksum>03F02484</xml_checksum>
   <class_name>BuildResidentialScheduleFile</class_name>
   <display_name>Schedule File Builder</display_name>
@@ -909,7 +909,7 @@
       <checksum>5FC694CF</checksum>
     </file>
     <file>
-      <filename>build_residential_schedule_file_test.rb</filename>
+      <filename>test_build_residential_schedule_file.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
       <checksum>BF7A1992</checksum>
