+++ ReportSimulationOutput/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>report_simulation_output</name>
   <uid>df9d170c-c21a-4130-866d-0d46b06073fd</uid>
-  <version_id>d0571607-acd4-408f-8345-edd99e077dda</version_id>
-  <version_modified>2023-11-01T12:23:52Z</version_modified>
+  <version_id>74e801f5-20a4-47bd-966e-d1378819545f</version_id>
+  <version_modified>2023-11-02T14:40:03Z</version_modified>
   <xml_checksum>9BF1E6AC</xml_checksum>
   <class_name>ReportSimulationOutput</class_name>
   <display_name>HPXML Simulation Output Report</display_name>
@@ -1932,7 +1932,7 @@
       <checksum>9C20F671</checksum>
     </file>
     <file>
-      <filename>output_report_test.rb</filename>
+      <filename>test_report_sim_output.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
       <checksum>CFAF52E1</checksum>
