+++ ReportUtilityBills/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>report_utility_bills</name>
   <uid>ca88a425-e59a-4bc4-af51-c7e7d1e960fe</uid>
-  <version_id>0051eed2-98c2-4c5e-b685-67436184aa78</version_id>
-  <version_modified>2023-10-31T19:02:55Z</version_modified>
+  <version_id>65fc31a4-1c57-4d29-9bd4-ef00e35c5c5f</version_id>
+  <version_modified>2023-11-02T14:40:05Z</version_modified>
   <xml_checksum>15BF4E57</xml_checksum>
   <class_name>ReportUtilityBills</class_name>
   <display_name>Utility Bills Report</display_name>
@@ -319,7 +319,7 @@
       <checksum>F96CB80F</checksum>
     </file>
     <file>
-      <filename>utility_bills_test.rb</filename>
+      <filename>test_report_utility_bills.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
       <checksum>217F2392</checksum>
