+++ tasks.rb
@@ -2366,21 +2366,21 @@ if ARGV[0].to_sym == :create_release_zips
 
   files = ['Changelog.md',
            'LICENSE.md',
-           'BuildResidentialHPXML/measure.*',
+           'BuildResidentialHPXML/*.*',
            'BuildResidentialHPXML/resources/**/*.*',
-           'BuildResidentialScheduleFile/measure.*',
+           'BuildResidentialScheduleFile/*.*',
            'BuildResidentialScheduleFile/resources/**/*.*',
-           'HPXMLtoOpenStudio/measure.*',
+           'HPXMLtoOpenStudio/*.*',
            'HPXMLtoOpenStudio/resources/**/*.*',
-           'ReportSimulationOutput/measure.*',
+           'ReportSimulationOutput/*.*',
            'ReportSimulationOutput/resources/**/*.*',
-           'ReportUtilityBills/measure.*',
+           'ReportUtilityBills/*.*',
            'ReportUtilityBills/resources/**/*.*',
            'weather/*.*',
            'workflow/*.*',
            'workflow/real_homes/*.xml',
            'workflow/sample_files/*.xml',
-           'workflow/tests/*test*.rb',
+           'workflow/tests/*.rb',
            'workflow/tests/**/*.xml',
            'workflow/tests/**/*.csv',
            'documentation/index.html',
